backend/pyproject.toml
[tool.poetry]
name = "backend"
version = "0.1.0"
description = "Lumiere Trading Platform."
authors = ["MrLumiere0 <hashidlouis@gmail.com>"]
license = "MIT"
readme = "README.md"
packages = []
package-mode = false

[tool.poetry.dependencies]
python = "^3.11" 
django = "^5.2.7"
djangorestframework = "^3.16.1"
psycopg2-binary = "^2.9.11"
django-environ = "^0.12.0"
pydantic = "^2.12.3"
djangorestframework-simplejwt = "^5.5.1"
passlib = "^1.7.4"
gunicorn = "^23.0.0"


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
